requirements.txt
PyMuPDF>=1.23.0,<1.25.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
Pillow>=9.3.0
openai>=1.40.0
tqdm>=4.66.0
google-generativeai>=0.7.0
elevenlabs>=1.7.0
moviepy==1.0.3
numpy<2.3.0
imageio-ffmpeg>=0.4.9
streamlit>=1.36.0
replicate>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.1
boto3>=1.28.0
